Nie raportuj błędu, jeśli vsnprintf zmieścił się idealnie w buforze
W razie wątpliwości, fragment dokumentacji:

The functions snprintf() and vsnprintf() do not write more than size
bytes (including the trailing '\0'). If the output was truncated due to
this limit then the return value is the number of characters (not
including the trailing '\0') which would have been written to the final
string if enough space had been available. Thus, a return value of size
or more means that the output was truncated. (See also below under
NOTES.)